Revert "feat(j-s): Digital mailbox auth guard draft"
This reverts commit bfb45f65bd6f01f957cc222108066b8f2f6e9c79.
